pyproject.toml
[project]
name = "core-coder-v5"
version = "1.0.0"
description = "A DevOps engineer focused on infrastructure, automation, and reliability.  Your expertise includes: - CI/CD pipeline design and optimization - Cloud infrastructure (AWS, Azure, GCP) - Container orches"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Generated by Agent Forge™", email = "agent-forge@example.com"}
]
license = {text = "Apache-2.0"}
keywords = ["langchain", "ai-agent", "groq", "cli", "automation", "devops", "agent-forge"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core LangChain dependencies (matching langchain-agent-base versions)
    "langchain>=1.0.8",
    "langgraph>=1.0.3",
    "langsmith>=0.4.43",
    "langchain-community>=0.4.1",
    "langchain-groq>=0.2.0",
    "langchain-ollama>=0.2.0",
    "groq>=0.11.0",
    "langchain-huggingface>=0.1.0",
    
    # Protocol system dependencies (required by src.protocol, src.server)
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "semver>=3.0.0",
    
    # Storage and retrieval
    "qdrant-client>=1.16.0",
    "langchain-qdrant>=1.1.0",
    "sentence-transformers>=2.2.2",
    
    # Document processing (required by src.rag WebBaseLoader)
    "beautifulsoup4>=4.14.2",
    "lxml>=4.9.0",
    
    # Environment & utilities
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

server = [
    # Uncomment to enable FastAPI server mode
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

full = [
    # All optional dependencies
    "beautifulsoup4>=4.14.2",
    "pymupdf>=1.24.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jupyter>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/agent-forge/core-coder-v5"
Repository = "https://github.com/agent-forge/core-coder-v5"
Documentation = "https://github.com/agent-forge/core-coder-v5#readme"
Issues = "https://github.com/agent-forge/core-coder-v5/issues"

[project.scripts]
core-coder-v5 = "cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# Generated by Agent Forge™ © 2026
# Professional CLI & API Agent Designer - Commercial Grade
# Based on langchain-agent-base production standards
